refactor: extract confidence computation to its own module (C9)
_compute_confidence and its helpers (_detect_evidence_tiers,
_determine_evidence_tier, _determine_confidence_level) were orchestration
logic living in diff_filtering, which only owns dedup/redundancy — following
the compare() flow required a cross-module hop.

Move them to a dedicated abicheck/confidence.py (depends only on
checker_policy/detectors/model, so no checker<->diff_filtering cycle).
checker and the two test modules import from confidence; the historical
_compute_confidence name is kept as an alias. diff_filtering shrinks ~190
lines. Behaviour-preserving.
